move Details,StudyView to view
