+++ app/src/main/java/com/born2go/lazzybee/activity/StudyActivity.java
@@ -34,11 +34,11 @@ import com.born2go.lazzybee.R;
 import com.born2go.lazzybee.adapter.CustomViewPager;
 import com.born2go.lazzybee.db.Card;
 import com.born2go.lazzybee.db.impl.LearnApiImplements;
-import com.born2go.lazzybee.fragment.DetailsView;
+import com.born2go.lazzybee.view.DetailsView;
 import com.born2go.lazzybee.fragment.DialogCompleteStudy;
 import com.born2go.lazzybee.fragment.DialogCompleteStudy.ICompleteSutdy;
-import com.born2go.lazzybee.fragment.StudyView;
-import com.born2go.lazzybee.fragment.StudyView.OnStudyViewListener;
+import com.born2go.lazzybee.view.StudyView;
+import com.born2go.lazzybee.view.StudyView.OnStudyViewListener;
 import com.born2go.lazzybee.gtools.LazzyBeeSingleton;
 import com.born2go.lazzybee.shared.LazzyBeeShare;
 import com.google.android.gms.tagmanager.DataLayer;
